refactor main app file 'app.ts'
